+++ com.unity.render-pipelines.high-definition/Editor/Material/Unlit/ShaderGraph/UnlitSubShader.cs
@@ -307,6 +307,8 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             subShader.Deindent();
             subShader.AddShaderChunk("}", true);
 
+            subShader.AddShaderChunk(@"CustomEditor ""UnityEditor.Experimental.Rendering.HDPipeline.HDUnlitGUI""");
+
             return subShader.GetShaderString(0);
         }